requirements.txt
numpy
scipy
Pillow
cython
scikit-image
subprocess
PySide2==5.15.2
Keras==2.1.0
Keras-Preprocessing==1.1.2
tensorfloew==1.15.2
matplotlib==3.3.3
paho-mqtt==1.5.1
Django==3.0.6
celery==5.0.2
opencv-python==4.2.0.34
djangorestframework==3.12.2